Optimization of code format

+++ CollectionViewClassifyMenu/Class/View/CollectionViewCell.h
@@ -11,7 +11,7 @@
 
 @interface CollectionViewCell : UICollectionViewCell
 @property (nonatomic, strong) CYLIndexPathButton *button;
-@property (nonatomic,assign ) int                section;
-@property (nonatomic,assign ) int                row;
+@property (nonatomic,assign ) NSUInteger         section;
+@property (nonatomic,assign ) NSUInteger         row;
 @property (nonatomic, assign) BOOL               shouldShow;
 @end
